# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Key Manager",
    "description": "Manage and validate API keys for 37+ AI providers. Import, check validity, test token limits and concurrency, query balances, and export working keys.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/logs": {
      "get": {
        "tags": [
          "Logs"
        ],
        "summary": "Api Logs",
        "description": "Get recent log entries.",
        "operationId": "api_logs_api_logs_get",
        "parameters": [
          {
            "name": "lines",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Number of recent log lines",
              "default": 100,
              "title": "Lines"
            },
            "description": "Number of recent log lines"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/logs/operations": {
      "get": {
        "tags": [
          "Logs"
        ],
        "summary": "Api Logs Operations",
        "description": "Get structured operations log.",
        "operationId": "api_logs_operations_api_logs_operations_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Max operations to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Max operations to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogEntry": {
        "properties": {
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp",
            "description": "Log entry timestamp"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level",
            "description": "Log level (INFO, WARNING, ERROR)"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Log message content",
            "default": ""
          },
          "extra": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra",
            "description": "Additional structured data"
          }
        },
        "type": "object",
        "title": "LogEntry"
      },
      "LogsResponse": {
        "properties": {
          "logs": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LogEntry"
                },
                {
                  "type": "string"
                }
              ]
            },
            "type": "array",
            "title": "Logs",
            "description": "Recent log entries"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Number of entries returned",
            "default": 0
          }
        },
        "type": "object",
        "title": "LogsResponse"
      },
...
      "OperationEntry": {
        "properties": {
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp",
            "description": "Operation timestamp"
          },
          "action": {
            "type": "string",
            "title": "Action",
            "description": "Operation type (import, check, test, etc.)",
            "default": ""
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Operation detail",
            "default": ""
          },
          "extra": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra",
            "description": "Additional structured data"
          }
        },
        "type": "object",
        "title": "OperationEntry"
      },
      "OperationsResponse": {
        "properties": {
          "operations": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OperationEntry"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                }
              ]
            },
            "type": "array",
            "title": "Operations",
            "description": "Structured operations log"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Number of entries returned",
            "default": 0
          }
        },
        "type": "object",
        "title": "OperationsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}